+++ src/main/java/com/github/kmu_shell_we/domain/season/_team/_team_mission/_submission/controller/SubmissionController.java
@@ -1,13 +1,14 @@
-package com.github.kmu_shell_we.domain.season._team._mission_team._submission_mission_team.controller;
+package com.github.kmu_shell_we.domain.season._team._team_mission._submission.controller;
 
 
-import com.github.kmu_shell_we.domain.mission.entity.Mission;
-import com.github.kmu_shell_we.domain.season._team._mission_team._submission_mission_team.dto.request.CreateSubmissionRequest;
-import com.github.kmu_shell_we.domain.season._team._mission_team._submission_mission_team.dto.response.SubmissionListResponse;
-import com.github.kmu_shell_we.domain.season._team._mission_team._submission_mission_team.dto.response.SubmissionResponse;
-import com.github.kmu_shell_we.domain.season._team._mission_team._submission_mission_team.service.SubmissionService;
+import com.github.kmu_shell_we.domain.season._team._team_mission._submission.dto.request.CreateSubmissionRequest;
+import com.github.kmu_shell_we.domain.season._team._team_mission._submission.dto.response.SubmissionListResponse;
+import com.github.kmu_shell_we.domain.season._team._team_mission._submission.dto.response.SubmissionResponse;
+import com.github.kmu_shell_we.domain.season._team._team_mission._submission.service.SubmissionService;
+import com.github.kmu_shell_we.domain.season._team._team_mission.entity.TeamMission;
 import com.github.kmu_shell_we.domain.season._team.entity.Team;
 import com.github.kmu_shell_we.domain.season.entity.Season;
+import com.github.kmu_shell_we.domain.user.entity.User;
 import com.github.kmu_shell_we.global.response.ApiResponse;
 import com.github.kmu_shell_we.global.security.guard.MemberGuard;
 import io.swagger.v3.oas.annotations.Operation;
@@ -16,6 +17,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @MemberGuard
@@ -30,33 +32,36 @@ public class SubmissionController {
     @GetMapping("/submissions")
     @Operation(summary = "팀 미션 제출 목록 조회")
     public ApiResponse<SubmissionListResponse> getSubmissions(
+            @AuthenticationPrincipal User user,
             @Parameter(description = "시즌 ID", schema = @Schema(type = "string", format = "uuid")) @PathVariable Season season,
             @Parameter(description = "팀 ID", schema = @Schema(type = "string", format = "uuid")) @PathVariable Team team
     ) {
 
-        return ApiResponse.ok(submissionService.getSubmissions(season, team));
+        return ApiResponse.ok(submissionService.getSubmissions(user, season, team));
     }
 
-    @GetMapping("/missions/{mission}/submissions")
+    @GetMapping("/team_mission/{team_mission}/submissions")
     @Operation(summary = "팀 미션 제출 조회")
     public ApiResponse<SubmissionResponse> getSubmission(
+            @AuthenticationPrincipal User user,
             @Parameter(description = "시즌 ID", schema = @Schema(type = "string", format = "uuid")) @PathVariable Season season,
             @Parameter(description = "팀 ID", schema = @Schema(type = "string", format = "uuid")) @PathVariable Team team,
-            @Parameter(description = "미션 ID", schema = @Schema(type = "string", format = "uuid"))@PathVariable Mission mission
+            @Parameter(description = "팀 미션 ID", schema = @Schema(type = "string", format = "uuid"))@PathVariable TeamMission team_mission
     ) {
 
-        return ApiResponse.ok(submissionService.getSubmission(season, team, mission));
+        return ApiResponse.ok(submissionService.getSubmission(user, season, team, team_mission));
     }
 
-    @PostMapping("/missions/{mission}/submissions")
+    @PostMapping("/team_mission/{team_mission}/submissions")
     @Operation(summary = "팀 미션 제출")
     public ApiResponse<SubmissionResponse> submitMission(
+            @AuthenticationPrincipal User user,
             @Parameter(description = "시즌 ID", schema = @Schema(type = "string", format = "uuid")) @PathVariable Season season,
             @Parameter(description = "팀 ID", schema = @Schema(type = "string", format = "uuid")) @PathVariable Team team,
-            @Parameter(description = "미션 ID", schema = @Schema(type = "string", format = "uuid")) @PathVariable Mission mission,
+            @Parameter(description = "팀 미션 ID", schema = @Schema(type = "string", format = "uuid")) @PathVariable TeamMission team_mission,
             @RequestBody @Valid CreateSubmissionRequest request
     ) {
 
-        return ApiResponse.ok(submissionService.submitSubmission(season, team, mission, request));
+        return ApiResponse.ok(submissionService.submitSubmission(user, season, team, team_mission, request));
     }
 }
