+++ src/main/java/com/github/kmu_shell_we/domain/season/_team/_team_mission/_submission/entity/Submission.java
@@ -1,6 +1,6 @@
-package com.github.kmu_shell_we.domain.season._team._mission_team._submission_mission_team.entity;
+package com.github.kmu_shell_we.domain.season._team._team_mission._submission.entity;
 
-import com.github.kmu_shell_we.domain.season._team._mission_team.entity.TeamMission;
+import com.github.kmu_shell_we.domain.season._team._team_mission.entity.TeamMission;
 import com.github.kmu_shell_we.global.infra.mysql.BaseSchema;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
