+++ src/main/java/com/github/kmu_shell_we/domain/season/_team/_team_mission/_submission/repository/SubmissionRepository.java
@@ -1,7 +1,7 @@
-package com.github.kmu_shell_we.domain.season._team._mission_team._submission_mission_team.repository;
+package com.github.kmu_shell_we.domain.season._team._team_mission._submission.repository;
 
-import com.github.kmu_shell_we.domain.season._team._mission_team._submission_mission_team.entity.Submission;
-import com.github.kmu_shell_we.domain.season._team._mission_team.entity.TeamMission;
+import com.github.kmu_shell_we.domain.season._team._team_mission._submission.entity.Submission;
+import com.github.kmu_shell_we.domain.season._team._team_mission.entity.TeamMission;
 
 import org.springframework.data.jpa.repository.JpaRepository;
