+++ src/main/java/com/github/kmu_shell_we/domain/season/_team/_team_mission/repository/TeamMissionRepository.java
@@ -1,10 +1,7 @@
-package com.github.kmu_shell_we.domain.season._team._mission_team.repository;
+package com.github.kmu_shell_we.domain.season._team._team_mission.repository;
 
-import com.github.kmu_shell_we.domain.mission.entity.Mission;
-import com.github.kmu_shell_we.domain.season._team._mission_team.entity.TeamMission;
+import com.github.kmu_shell_we.domain.season._team._team_mission.entity.TeamMission;
 import com.github.kmu_shell_we.domain.season._team.entity.Team;
-
-import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
@@ -14,9 +11,6 @@ import java.util.Optional;
 import java.util.UUID;
 
 public interface TeamMissionRepository extends JpaRepository<TeamMission, UUID> {
-
-    @EntityGraph(attributePaths = {"team", "mission"})
-    Optional<TeamMission> findByTeamAndMission(Team team, Mission mission);
 
     List<TeamMission> findAllByTeamAndEndedAtGreaterThanEqual(Team team, LocalDateTime now);
